requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
atari-py=0.2.6=pypi_0
atomicwrites=1.3.0=pypi_0
attrs=19.3.0=pypi_0
box2d-py=2.3.8=pypi_0
bzip2=1.0.8=h01d97ff_1
ca-certificates=2019.9.11=hecc5488_0
cairo=1.16.0=he1c11cd_1002
certifi=2019.9.11=py37_0
cffi=1.13.1=pypi_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
cloudpickle=1.2.2=pypi_0
colorama=0.4.1=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.13=py37h6de7cb9_0
decorator=4.4.0=pypi_0
ffmpeg=4.1.3=h5c2b479_0
fontconfig=2.13.1=h6b1039f_1001
freetype=2.10.0=h24853df_1
future=0.18.1=py37_0
gettext=0.19.8.1=h46ab8bc_1002
giflib=5.1.7=h01d97ff_1
glfw=1.8.3=pypi_0
glib=2.58.3=h9d45998_1002
gmp=6.1.2=h0a44026_1000
gnutls=3.6.5=h53004b3_1002
graphite2=1.3.13=h2098e52_1000
gym=0.14.0=dev_0
gym-minigrid=0.0.5=dev_0
h5py=2.10.0=pypi_0
harfbuzz=2.4.0=hd8d2a14_3
hdf5=1.10.5=nompi_h15a436c_1103
icu=64.2=h6de7cb9_1
idna=2.8=pypi_0
imageio=2.6.1=pypi_0
importlib-metadata=0.23=pypi_0
jasper=1.900.1=h636a363_1006
jedi=0.15.1=pypi_0
joblib=0.14.0=pypi_0
jpeg=9c=h1de35cc_1001
jsonpatch=1.24=pypi_0
jsonpointer=2.0=pypi_0
kaptan=0.5.12=pypi_0
kiwisolver=1.1.0=pypi_0
lame=3.100=h1de35cc_1001
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=4.0.0=2
libiconv=1.15=h01d97ff_1005
liblapack=3.8.0=14_openblas
liblapacke=3.8.0=14_openblas
libopenblas=0.3.7=h4bb4525_2
libpng=1.6.37=h2573ce8_0
libtiff=4.0.10=hd08fb8f_1003
libtmux=0.8.2=pypi_0
libwebp=1.0.2=h14d73da_2
libxml2=2.9.9=h12c6b28_5
llvm-openmp=9.0.0=h40edb58_0
lockfile=0.12.2=pypi_0
lz4-c=1.8.3=h6de7cb9_1001
matplotlib=3.1.1=pypi_0
more-itertools=7.2.0=pypi_0
mpi=1.0=openmpi
mpi4py=3.0.2=py37h5017f14_0
mujoco-py=2.0.2.5=dev_0
ncurses=6.1=h0a44026_1
nettle=3.4.1=h3efe00b_1002
networkx=2.4=pypi_0
numpy=1.17.3=pypi_0
olefile=0.46=py_0
opencv=3.4.7=py37hb60cc42_3
opencv-python=4.1.1.26=pypi_0
openh264=1.8.0=hd9629dc_1000
openmpi=4.0.1=hfcebdee_2
openssl=1.1.1c=h01d97ff_0
packaging=19.2=pypi_0
pandas=0.25.2=pypi_0
parso=0.5.1=pypi_0
pcre=8.43=h0a44026_0
pillow=6.2.1=pypi_0
pip=19.3.1=py37_0
pixman=0.38.0=h01d97ff_1003
pluggy=0.13.0=pypi_0
psutil=5.6.3=pypi_0
py=1.8.0=pypi_0
pycparser=2.19=pypi_0
pyglet=1.3.2=pypi_0
pyopengl=3.1.3b2=py_1
pyparsing=2.4.2=pypi_0
pyqt5=5.13.1=pypi_0
pyqt5-sip=12.7.0=pypi_0
pytest=5.2.1=pypi_0
pytest-instafail=0.4.1=pypi_0
python=3.7.4=h359304d_1
python-dateutil=2.8.0=pypi_0
python-jsonrpc-server=0.2.0=pypi_0
pytz=2019.3=pypi_0
pyvips=2.1.8=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.1.2=pypi_0
pyzmq=18.1.0=pypi_0
readline=7.0=h1de35cc_5
requests=2.22.0=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.1=pypi_0
setuptools=41.4.0=py37_0
six=1.12.0=pypi_0
sqlite=3.30.0=ha441bb4_0
tk=8.6.8=ha441bb4_0
tmuxp=1.5.3=pypi_0
torch=1.3.0=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.4.1=pypi_0
tornado=6.0.3=pypi_0
tqdm=4.36.1=pypi_0
urllib3=1.25.6=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.1.7=pypi_0
websocket-client=0.56.0=pypi_0
wheel=0.33.6=py37_0
wrapt=1.11.2=pypi_0
x264=1!152.20180806=h1de35cc_0
xz=5.2.4=h1de35cc_4
zipp=0.6.0=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.0=ha9f0a20_0
